[CHORE] Extract Bud Box schedule utility
